+++ src/strategies/interfaces/ILiquidityProvider.sol
@@ -22,36 +22,20 @@ interface ILiquidityProvider is IOfferLogic {
   ///@param inbound_tkn the outbound token of the offer list of the offer
   ///@param wants the amount of outbound tokens the offer maker requires for a complete fill
   ///@param gives the amount of inbound tokens the offer maker gives for a complete fill
-  ///@param gasreq the amount of gas units that are required to execute the trade (use type(uint).max for using `this.offerGasreq()`)
-  ///@param gasprice the gasprice used to compute offer's provision (use 0 to use Mangrove's gasprice)
   ///@param pivotId the pivot to use for inserting the offer in the list
-  function newOffer(
-    IERC20 outbound_tkn,
-    IERC20 inbound_tkn,
-    uint wants,
-    uint gives,
-    uint gasreq,
-    uint gasprice,
-    uint pivotId
-  ) external payable returns (uint);
+  function newOffer(IERC20 outbound_tkn, IERC20 inbound_tkn, uint wants, uint gives, uint pivotId)
+    external
+    payable
+    returns (uint);
 
   ///@notice updates an offer existing on Mangrove (not necessarily live).
   ///@param outbound_tkn the outbound token of the offer list of the offer
   ///@param inbound_tkn the outbound token of the offer list of the offer
   ///@param wants the new amount of outbound tokens the offer maker requires for a complete fill
   ///@param gives the new amount of inbound tokens the offer maker gives for a complete fill
-  ///@param gasreq the new amount of gas units that are required to execute the trade (use type(uint).max for using `this.offerGasReq()`)
-  ///@param gasprice the new gasprice used to compute offer's provision (use 0 to use Mangrove's gasprice)
   ///@param pivotId the pivot to use for re-inserting the offer in the list (use `offerId` if updated offer is live)
   ///@param offerId the id of the offer in the offer list.
-  function updateOffer(
-    IERC20 outbound_tkn,
-    IERC20 inbound_tkn,
-    uint wants,
-    uint gives,
-    uint gasreq,
-    uint gasprice,
-    uint pivotId,
-    uint offerId
-  ) external payable;
+  function updateOffer(IERC20 outbound_tkn, IERC20 inbound_tkn, uint wants, uint gives, uint pivotId, uint offerId)
+    external
+    payable;
 }